// src/app/Entity/Model/tbl_stok_karti.ts
export class tbl_stok_karti {

    rec_id: number;
    rec_status: number;
    rec_date_added: Date;
    rec_user_added: number;
    rec_date_updated: Date;
    rec_user_updated: number;
    stk_kod: string;
    stk_adi: string;
    stk_tip: string;
    stk_birim_ref: number;
    stk_ozel_kod_1: string;
    stk_ozel_kod_2: string;
    stk_ozel_kod_3: string;
    stk_ozel_kod_4: string;
    stk_ozel_kod_5: string;
    stk_alis_fiyati_1: number;
    stk_alis_fiyati_2: number;
    stk_satis_fiyati_1: number;
    stk_satis_fiyati_2: number;
    stk_satis_fiyati_3: number;
    stk_alis_kdv_orani: number;
    stk_satis_kdv_orani: number;
    stk_otv_kod: string;
    stk_otv_orani: number;
    stk_en_az_miktar: number;
    stk_en_cok_miktar: number;
    stk_optimum_miktar: number;

}